+++ packages/0xsequence/tests/browser/window-transport/dapp.test.ts
@@ -9,7 +9,7 @@ import { configureLogger, encodeMessageDigest, packMessageData } from '@0xsequen
 
 import { testWalletContext } from '../testutils'
 
-configureLogger({ logLevel: 'DEBUG' })
+configureLogger({ logLevel: 'DEBUG', silence: false })
 
 const walletProvider = new WindowMessageProvider('http://localhost:9999/mock-wallet/mock-wallet.test.html')
 walletProvider.register()