+++ packages/network/src/json-rpc/middleware/logging-provider.ts
@@ -1,15 +1,16 @@
 import { JsonRpcHandlerFunc, JsonRpcRequest, JsonRpcResponse, JsonRpcResponseCallback, JsonRpcMiddleware } from '../types'
+import { logger } from '@0xsequence/utils'
 
 export const loggingProviderMiddleware: JsonRpcMiddleware = (next: JsonRpcHandlerFunc) => {
   return (request: JsonRpcRequest, callback: JsonRpcResponseCallback, chainId?: number) => {
     const chainIdLabel = chainId ? ` chainId:${chainId}` : ''
-    console.log(`[provider request]${chainIdLabel} id:${request.id} method:${request.method} params:`, request.params)
+    logger.info(`[provider request]${chainIdLabel} id:${request.id} method:${request.method} params:`, request.params)
 
     next(request, (error: any, response?: JsonRpcResponse) => {
       if (error) {
-        console.warn(`[provider response]${chainIdLabel} id:${request.id} method:${request.method} params:`, request.params, `error:`, error)
+        logger.warn(`[provider response]${chainIdLabel} id:${request.id} method:${request.method} params:`, request.params, `error:`, error)
       } else {
-        console.log(`[provider response]${chainIdLabel} id:${request.id} method:${request.method} params:`, request.params, `response:`, response)
+        logger.info(`[provider response]${chainIdLabel} id:${request.id} method:${request.method} params:`, request.params, `response:`, response)
       }
       callback(error, response)
     }, chainId)