+++ packages/multicall/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @0xsequence/multicall
 
+## 0.36.9
+
+### Patch Changes
+
+- minor dev xp improvements
+- Updated dependencies
+  - @0xsequence/abi@0.36.9
+  - @0xsequence/network@0.36.9
+  - @0xsequence/utils@0.36.9
+
 ## 0.36.8
 
 ### Patch Changes